refactor(tools): extract tool output policy
