+++ src/file-toolkit.ts
@@ -4,19 +4,14 @@ import { deleteTextFile, editFile, findFiles, readFileContents, searchFiles, wri
 import { createTool, type ToolkitDeps, type ToolkitInput } from "./tool-contract";
 import { runTool } from "./tool-execution";
 import { compactToolOutput } from "./tool-output";
-import {
-  diffSummaryParts,
-  emitParts,
-  findSummaryParts,
-  searchSummaryParts,
-  TOOL_PROGRESS_LIMITS,
-} from "./tool-output-format";
+import { diffSummaryParts, emitParts, findSummaryParts, searchSummaryParts } from "./tool-output-format";
 import {
   findResultPaths,
   numberedUnifiedDiffLines,
   searchResultSummaryStats,
   summarizeUnifiedDiff,
 } from "./tool-output-parse";
+import { TOOL_PROGRESS_LIMITS } from "./tool-policy";
 
 function normalizeUniquePaths(paths: string[]): string[] {
   const normalized = paths.map((path) => path.trim()).filter((path) => path.length > 0);
